add checks to prevent errors before update

+++ adm_themes/modern/my_body_bottom.php
@@ -31,11 +31,11 @@
     </a><br />
     <span style="font-size: 9pt; padding-left: 8px;">&copy; 2004 - 2018&nbsp;&nbsp;
         <?php echo $gL10n->get('SYS_ADMIDIO_TEAM'). '<br />';
-            if(strlen($gSettingsManager->getString('system_url_data_protection')) > 0)
+            if ($gSettingsManager->has('system_url_data_protection') && strlen($gSettingsManager->getString('system_url_data_protection')) > 0)
             {
                 echo '<a href="'.$gSettingsManager->getString('system_url_data_protection').'">'.$gL10n->get('SYS_DATA_PROTECTION').'</a>';
             }
-            if(strlen($gSettingsManager->getString('system_url_imprint')) > 0)
+            if ($gSettingsManager->has('system_url_imprint') && strlen($gSettingsManager->getString('system_url_imprint')) > 0)
             {
                 echo '&nbsp;&nbsp;&nbsp;<a href="'.$gSettingsManager->getString('system_url_imprint').'">'.$gL10n->get('SYS_IMPRINT').'</a>';
             }
